[TRAV-85] - implementing post review comments
