--- cheddar/cheddar-integration-aws/src/main/java/com/clicktravel/infrastructure/persistence/aws/dynamodb/manager/DynamoDocumentStoreTemplateInfrastructureManager.java
@@ -1,53 +1,0 @@
-/*
- * Copyright 2014 Click Travel Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-package com.clicktravel.infrastructure.persistence.aws.dynamodb.manager;
-
-import java.util.Collection;
-import java.util.HashSet;
-
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
-import com.clicktravel.infrastructure.persistence.aws.dynamodb.DynamoDocumentStoreTemplate;
-
-/**
- * Amazon Web Services Persistence Infrastructure Manager
- *
- * The responsibility of implementing classes is to create tables
- */
-public class DynamoDocumentStoreTemplateInfrastructureManager extends AbstractPersistenceManager {
-
-    private final Collection<DynamoDocumentStoreTemplate> dynamoDocumentStoreTemplates;
-
-    public DynamoDocumentStoreTemplateInfrastructureManager(final AmazonDynamoDB amazonDynamoDbClient,
-            final int readThroughput, final int writeThroughput) {
-        super(amazonDynamoDbClient, readThroughput, writeThroughput);
-        dynamoDocumentStoreTemplates = new HashSet<DynamoDocumentStoreTemplate>();
-    }
-
-    public void setDynamoDocumentStoreTemplates(
-            final Collection<DynamoDocumentStoreTemplate> dynamoDocumentStoreTemplates) {
-        if (dynamoDocumentStoreTemplates != null) {
-            this.dynamoDocumentStoreTemplates.addAll(dynamoDocumentStoreTemplates);
-        }
-    }
-
-    public void init() {
-        for (final DynamoDocumentStoreTemplate dynamoDocumentStoreTemplate : dynamoDocumentStoreTemplates) {
-            super.initSchema(dynamoDocumentStoreTemplate.databaseSchemaHolder());
-            dynamoDocumentStoreTemplate.initialize(amazonDynamoDbClient);
-        }
-    }
-}
